some drive examples
