+++ examples/tankdrive-robot/.gitignore
@@ -1,0 +1,3 @@
+/.gradle/
+/build/
+/.idea/
