+++ examples/tankdrive-robot/src/main/java/frc/team3388/robot/Main.java
@@ -1,0 +1,10 @@
+package frc.team3388.robot;
+
+import com.flash3388.flashlib.frc.robot.RobotMain;
+
+public class Main {
+
+    public static void main(String[] args) {
+        RobotMain.start(Robot::new);
+    }
+}
