+++ examples/tankdrive-robot/src/main/java/frc/team3388/robot/Robot.java
@@ -1,0 +1,139 @@
+package frc.team3388.robot;
+
+import com.flash3388.flashlib.frc.robot.base.iterative.IterativeFrcRobot;
+import com.flash3388.flashlib.frc.robot.hid.JoystickAxis;
+import com.flash3388.flashlib.frc.robot.systems.Systems;
+import com.flash3388.flashlib.hid.Hid;
+import com.flash3388.flashlib.robot.RobotControl;
+import com.flash3388.flashlib.robot.base.DelegatingRobotControl;
+import com.flash3388.flashlib.robot.systems.drive.TankDriveSystem;
+import com.flash3388.flashlib.robot.systems.drive.actions.ArcadeDriveAction;
+import com.flash3388.flashlib.robot.systems.drive.actions.TankDriveAction;
+import com.flash3388.flashlib.scheduling.actions.Actions;
+import edu.wpi.first.wpilibj.PWMTalonSRX;
+
+public class Robot extends DelegatingRobotControl implements IterativeFrcRobot {
+
+    private final TankDriveSystem mDriveSystem;
+
+    private final Hid mStickRight;
+    private final Hid mStickLeft;
+
+    public Robot(RobotControl robotControl) {
+        super(robotControl);
+
+        // Creating the tank drive system.
+        // We define the speed controllers for each side, there are 2 per each.
+        mDriveSystem = Systems.newTankDrive()
+                .right(new PWMTalonSRX(RobotMap.DRIVE_FRONT_RIGHT))
+                .right(new PWMTalonSRX(RobotMap.DRIVE_BACK_RIGHT))
+                .left(new PWMTalonSRX(RobotMap.DRIVE_FRONT_LEFT))
+                .left(new PWMTalonSRX(RobotMap.DRIVE_BACK_LEFT))
+                .build();
+
+        // Creating the joysticks.
+        // We will use these joysticks to control the motions of the drive system.
+        mStickRight = getHidInterface().newGenericHid(RobotMap.STICK_RIGHT);
+        mStickLeft = getHidInterface().newGenericHid(RobotMap.STICK_LEFT);
+
+        // We set the default action for the drive system.
+        // This action, will run by default on the system.
+        // Using it, we can define the default behaviour for the system.
+        // In this case, the default behaviour is to move the system as such:
+        //      - Right joystick - Y axis: move right side of drive
+        //      - Left joystick - Y axis: move left side of drive
+        //
+        // Note the `requires` in the end. This is superrrrr important.
+        // It defines to the Scheduler (which runs the actions) that mDriveSystem is used
+        // in this action. Using this, the Scheduler will prevent two actions from using the
+        // drive system at the same time.
+        mDriveSystem.setDefaultAction(new TankDriveAction(mDriveSystem,
+                mStickRight.getAxis(JoystickAxis.Y),
+                mStickLeft.getAxis(JoystickAxis.Y))
+                .requires(mDriveSystem));
+
+        // We can define actions to be executed when a button is pressed.
+        // Here we define that when button number 0 (labeled on the joystick) is pressed,
+        // the joystick axes are inverted, making forwards backwards, and backwards forwards.
+        //
+        // `Actions.instant` allows defining an actions which only runs once. The actual content
+        // of the action, is defined in the lambda:
+        // ()-> {
+        //      // do stuff in action
+        // }
+        //
+        // Because we defined that by default, the drive system runs an action that moves using
+        // the joysticks; when this is pressed, we will have reverse motion from the drive system.
+        mStickLeft.getButton(0).whenActive(Actions.instant(()-> {
+            mStickRight.getAxis(JoystickAxis.Y).invert();
+            mStickLeft.getAxis(JoystickAxis.Y).invert();
+        }));
+
+        // Like before, we are creating another action to run when a button is pressed.
+        // This time, when the button is pressed, one of two possibilities can occur:
+        //      - if not already running: a non-instant action will run, which will move the drive system
+        //        based on the joysticks, but in an arcade-drive algorithm instead of tank-drive.
+        //      - if already running: the mentioned action will be canceled, and the default
+        //        action will start running.
+        //
+        // In essence, this button will toggle the action, switching between the default and it.
+        // Or in other words: switching between tank-drive and arcade-drive.
+        //
+        // Notice the use of `requires` again. Thanks to this, when this action runs, the Scheduler will stop
+        // the default action, thus allowing only one of them to run at any given moment.
+        mStickLeft.getButton(1).toggleWhenActive(new ArcadeDriveAction(mDriveSystem,
+                mStickRight.getAxis(JoystickAxis.Y),
+                mStickLeft.getAxis(JoystickAxis.Y))
+                .requires(mDriveSystem));
+    }
+
+    @Override
+    public void disabledInit() {
+
+    }
+
+    @Override
+    public void disabledPeriodic() {
+
+    }
+
+    @Override
+    public void teleopInit() {
+
+    }
+
+    @Override
+    public void teleopPeriodic() {
+
+    }
+
+    @Override
+    public void autonomousInit() {
+
+    }
+
+    @Override
+    public void autonomousPeriodic() {
+
+    }
+
+    @Override
+    public void testInit() {
+
+    }
+
+    @Override
+    public void testPeriodic() {
+
+    }
+
+    @Override
+    public void robotPeriodic() {
+
+    }
+
+    @Override
+    public void robotStop() {
+
+    }
+}
