+++ examples/tankdrive-robot/src/main/java/frc/team3388/robot/RobotMap.java
@@ -9,7 +9,6 @@ public class RobotMap {
     public static final int DRIVE_BACK_RIGHT = 1;
     public static final int DRIVE_FRONT_LEFT = 2;
     public static final int DRIVE_BACK_LEFT = 3;
-    public static final int SHOOTER = 4;
 
     public static final HidChannel STICK_RIGHT = RoboRio.newHidChannel(0);
     public static final HidChannel STICK_LEFT = RoboRio.newHidChannel(1);
